requirements.txt
gTTS
discord.py[voice]
openai
tiktoken